Refactoring
Trying to rename the folders.
